Last console version of my Finance, Budget, Expense
The next version will be using a GUI and in NetBeans.  I had issues creating the GUI in BlueJay and it was time to move to a more powerful IDE anyway.  This version is fully functional with SQLite.  of course you need to set up Sqlite and the folders to make the application work.  V02 is still all just in memory.